refactor(suggestion): implement async request manager for better abortion control
